Resolve viewport harness policy findings
