+++ Packages/iOS/CmuxMobileTerminal/Sources/CmuxMobileTerminal/TerminalViewportCoordinator.swift
@@ -54,7 +54,7 @@ struct TerminalViewportCoordinator {
             width: bounds.width,
             height: max(1, containerSize.height)
         )
-        let liveViewportHeight = Self.liveViewportHeight(
+        let liveViewportHeight = terminalLiveViewportHeight(
             inputs: inputs,
             boundsHeight: bounds.height,
             fallbackHeight: layoutViewport.height
@@ -75,71 +75,19 @@ struct TerminalViewportCoordinator {
         )
     }
 
-    private static func liveViewportHeight(
-        inputs: TerminalViewportInputs,
-        boundsHeight: CGFloat,
-        fallbackHeight: CGFloat
-    ) -> CGFloat {
-        guard inputs.chromeVisible,
-              let frame = inputs.toolbarPresentationFrame ?? inputs.toolbarFrame,
-              !frame.isNull,
-              !frame.isEmpty else {
-            return fallbackHeight
-        }
-        return min(max(1, frame.minY), max(1, boundsHeight))
-    }
 }
 
-struct TerminalViewportInputs {
-    let bounds: CGSize
-    let keyboardHeight: CGFloat
-    let composerBandHeight: CGFloat
-    let reservedToolbarHeight: CGFloat
-    let toolbarFrameHeight: CGFloat
-    let bottomSafeAreaInset: CGFloat
-    let chromeHidden: Bool
-    let chromeVisible: Bool
-    let toolbarFrame: CGRect?
-    let toolbarPresentationFrame: CGRect?
-}
-
-struct TerminalViewportSnapshot {
-    let bounds: CGSize
-    let containerSize: CGSize
-    let keyboardOccupancy: CGFloat
-    let composerFrame: CGRect
-    let toolbarFrame: CGRect
-    let layoutViewportRect: CGRect
-    let liveViewportRect: CGRect
-
-    func renderViewportRect(forRenderSize renderSize: CGSize, clampsStaleLiveViewport: Bool) -> CGRect {
-        let targetHeight = layoutViewportRect.height
-        let liveHeight = liveViewportRect.height
-        let height = clampsStaleLiveViewport ? min(liveHeight, targetHeight) : liveHeight
-        return CGRect(
-            x: layoutViewportRect.minX,
-            y: layoutViewportRect.minY,
-            width: layoutViewportRect.width,
-            height: max(1, height)
-        )
+private func terminalLiveViewportHeight(
+    inputs: TerminalViewportInputs,
+    boundsHeight: CGFloat,
+    fallbackHeight: CGFloat
+) -> CGFloat {
+    guard inputs.chromeVisible,
+          let frame = inputs.toolbarPresentationFrame ?? inputs.toolbarFrame,
+          !frame.isNull,
+          !frame.isEmpty else {
+        return fallbackHeight
     }
-
-    func renderRect(forRenderSize renderSize: CGSize, clampsStaleLiveViewport: Bool) -> CGRect {
-        let viewport = renderViewportRect(
-            forRenderSize: renderSize,
-            clampsStaleLiveViewport: clampsStaleLiveViewport
-        )
-        return CGRect(
-            x: viewport.minX,
-            y: viewport.maxY - renderSize.height,
-            width: renderSize.width,
-            height: renderSize.height
-        )
-    }
-
-    func isLetterboxed(renderSize: CGSize) -> Bool {
-        renderSize.width + 0.5 < layoutViewportRect.width
-            || renderSize.height + 0.5 < layoutViewportRect.height
-    }
+    return min(max(1, frame.minY), max(1, boundsHeight))
 }
 #endif
